+++ packages/gamut/src/GridForm/GridFormInputGroup/index.tsx
@@ -4,13 +4,18 @@ import { UseFormMethods } from 'react-hook-form';
 import { FormError, FormGroup, FormGroupLabel } from '../../Form';
 import { HiddenText } from '../../HiddenText';
 import { Column } from '../../Layout';
-import { GridFormField } from '../types';
+import {
+  GridFormField,
+  GridFormHiddenField,
+  GridFormSweetContainerField,
+} from '../types';
 import { GridFormCheckboxInput } from './GridFormCheckboxInput';
 import { GridFormCustomInput } from './GridFormCustomInput';
 import { GridFormFileInput } from './GridFormFileInput';
 import { GridFormHiddenInput } from './GridFormHiddenInput';
 import { GridFormRadioGroupInput } from './GridFormRadioGroupInput';
 import { GridFormSelectInput } from './GridFormSelectInput';
+import { GridFormSweetContainerInput } from './GridFormSweetContainerInput';
 import { GridFormTextArea } from './GridFormTextArea';
 import { GridFormTextInput } from './GridFormTextInput';
 
@@ -48,6 +53,7 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
         );
 
       case 'custom':
+      case 'custom-group':
         return (
           <GridFormCustomInput
             field={field}
@@ -99,6 +105,15 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'hidden':
         return <GridFormHiddenInput register={register} field={field} />;
 
+      case 'sweet-container':
+        return (
+          <GridFormSweetContainerInput
+            register={register}
+            field={field}
+            label={field.label}
+          />
+        );
+
       default:
         return (
           <GridFormTextInput
@@ -110,7 +125,23 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
         );
     }
   };
-  if (field.type === 'hidden') return getInput();
+
+  const unwrappedInput = (
+    field: GridFormField
+  ): field is GridFormHiddenField | GridFormSweetContainerField =>
+    ['hidden', 'sweet-container'].includes(field.type);
+
+  if (unwrappedInput(field)) {
+    return getInput();
+  }
+
+  if (field.type === 'custom-group') {
+    return (
+      <Column size={field?.size} rowspan={field?.rowspan ?? 1}>
+        {getInput()}
+      </Column>
+    );
+  }
 
   const label = (
     <FormGroupLabel
@@ -124,7 +155,7 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
   );
 
   return (
-    <Column size={field.size}>
+    <Column size={field?.size} rowspan={field?.rowspan ?? 1}>
       <FormGroup mb={0}>
         {field.hideLabel ? <HiddenText>{label}</HiddenText> : label}
         {getInput()}