+++ packages/gamut/src/ProLabel/index.tsx
@@ -1,6 +1,11 @@
-import { styledConfig, system } from '@codecademy/gamut-styles';
+import {
+  ColorModes,
+  styledConfig,
+  system,
+  useCurrentMode,
+  variant,
+} from '@codecademy/gamut-styles';
 import { StyleProps, variance } from '@codecademy/variance';
-import { Theme, useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 import React, { SVGProps } from 'react';
 
@@ -8,7 +13,7 @@ export type ProLabelProps = SVGProps<SVGSVGElement> &
   StyleProps<typeof logoStyles> &
   StyleProps<typeof placementVariants> &
   StyleProps<typeof modeVariants> & {
-    mode?: keyof Theme['colorModes']['modes'];
+    mode?: ColorModes;
     height?: number;
   };
 
@@ -18,7 +23,7 @@ const logoStyles = variance.compose(
   system.space
 );
 
-const placementVariants = system.variant({
+const placementVariants = variant({
   prop: 'placement',
   variants: {
     inline: {
@@ -27,7 +32,7 @@ const placementVariants = system.variant({
   },
 });
 
-const modeVariants = system.variant({
+const modeVariants = variant({
   prop: 'mode',
   variants: {
     light: { bg: 'navy', textColor: 'beige' },
@@ -46,14 +51,12 @@ export const ProLabel: React.FC<React.ComponentProps<typeof Svg>> = ({
   height = 16,
   ...props
 }) => {
-  const {
-    colorModes: { active },
-  } = useTheme();
+  const active = useCurrentMode(mode);
   return (
     <Svg
       {...props}
       height={height}
-      mode={mode ?? active}
+      mode={active}
       viewBox="0 0 30 16"
       xmlns="http://www.w3.org/2000/svg"
     >